refactor(mobile): analytics types
